==> main/06_Scheid_gemuteerd_en_ongemuteerd.R <==
## Author: Bastiaen Boekelo
## Date: July 2018

## Goal: Selecteer welke percelen gemuteerd zijn en welke niet


#####################################################
####### Determine with the selecte ########
#####################################################

## Keer terug naar Yearstats_all.csv

## Maak een selectie op basis van de combinatie die goed was o.b.v. script 5

## Merge de data met COV_and_difmax.csv -> Bepaal welke percelen er wellicht extra afvallen (op basis van threshold cov)




## ALTERNATIEF --> IMPLEMENTEER de COV threshold ook in script 5 en bepaal bij welke threshold je hoogste TN gehalte is bij een FN = 0.
